Refactored the embedding_generator
