+++ message_consumer/callbacks.py
@@ -1,5 +1,5 @@
 import logging
-from embedding_generator.task import process_images
+from embedding_generator.service.image_processing_service import ImageProcessingService
 
 logging.basicConfig(level=logging.INFO)
 
@@ -8,5 +8,6 @@ logger = logging.getLogger(__name__)
 
 def custom_callback(message):
     logger.info(f"Processing message from Kafka:{message}")
-    process_images(message)
+    service = ImageProcessingService()
+    service.process_images(message)
